agent-2-suno-prompt-generator/requirements.txt
google-cloud-aiplatform[adk,agent_engines]>=1.111
google-adk>=0.1.0
google-genai>=0.3.0
cloudpickle>=3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0
